chore: move logics to FilterSegmentRepository
