+++ packages/lib/server/repository/__tests__/filterSegments/delete.test.ts
@@ -1,16 +1,17 @@
-import prismock from "../../../../../../../tests/libs/__mocks__/prisma";
+import prismock from "../../../../../../tests/libs/__mocks__/prisma";
 
 import { describe, expect, it } from "vitest";
 
 import { MembershipRole } from "@calcom/prisma/enums";
 import type { TrpcSessionUser } from "@calcom/trpc/server/trpc";
 
-import { deleteHandler } from "../delete.handler";
-import { type TDeleteFilterSegmentInputSchema } from "../delete.schema";
+import { FilterSegmentRepository } from "../../filterSegment";
+import { type TDeleteFilterSegmentInputSchema } from "../../filterSegment.type";
 
-describe("deleteHandler", () => {
+describe("FilterSegmentRepository.delete()", () => {
+  const userId = 1;
   const mockUser = {
-    id: 1,
+    id: userId,
     name: "Test User",
   } as NonNullable<TrpcSessionUser>;
 
@@ -34,15 +35,12 @@ describe("deleteHandler", () => {
       id: segment.id,
     };
 
-    const result = await deleteHandler({
-      ctx: { user: mockUser },
-      input,
-    });
-
-    expect(result).toEqual({
-      id: segment.id,
-      message: "Filter segment deleted successfully",
-    });
+    await expect(
+      FilterSegmentRepository.delete({
+        userId,
+        input,
+      })
+    ).resolves.not.toThrow();
 
     // Verify segment is deleted
     const deletedSegment = await prismock.filterSegment.findUnique({
@@ -90,15 +88,12 @@ describe("deleteHandler", () => {
       id: segment.id,
     };
 
-    const result = await deleteHandler({
-      ctx: { user: mockUser },
-      input,
-    });
-
-    expect(result).toEqual({
-      id: segment.id,
-      message: "Filter segment deleted successfully",
-    });
+    await expect(
+      FilterSegmentRepository.delete({
+        userId,
+        input,
+      })
+    ).resolves.not.toThrow();
 
     // Verify segment is deleted
     const deletedSegment = await prismock.filterSegment.findUnique({
@@ -113,8 +108,8 @@ describe("deleteHandler", () => {
     };
 
     await expect(
-      deleteHandler({
-        ctx: { user: mockUser },
+      FilterSegmentRepository.delete({
+        userId,
         input,
       })
     ).rejects.toThrow("Filter segment not found or you don't have permission to delete it");
@@ -160,8 +155,8 @@ describe("deleteHandler", () => {
     };
 
     await expect(
-      deleteHandler({
-        ctx: { user: mockUser },
+      FilterSegmentRepository.delete({
+        userId,
         input,
       })
     ).rejects.toThrow("Filter segment not found or you don't have permission to delete it");
