--- packages/trpc/server/routers/viewer/filterSegments/delete.schema.ts
@@ -1,7 +1,0 @@
-import { z } from "zod";
-
-export const ZDeleteFilterSegmentInputSchema = z.object({
-  id: z.number().int().positive(),
-});
-
-export type TDeleteFilterSegmentInputSchema = z.infer<typeof ZDeleteFilterSegmentInputSchema>;
