+++ packages/trpc/server/routers/viewer/filterSegments/list.handler.ts
@@ -1,14 +1,6 @@
-import { prisma } from "@calcom/prisma";
+import { FilterSegmentRepository } from "@calcom/lib/server/repository/filterSegment";
+import type { TListFilterSegmentsInputSchema } from "@calcom/lib/server/repository/filterSegment.type";
 import type { TrpcSessionUser } from "@calcom/trpc/server/types";
-
-import type { TListFilterSegmentsInputSchema } from "./list.schema";
-import {
-  ZActiveFilters,
-  ZSortingState,
-  ZColumnSizing,
-  ZColumnVisibility,
-  type FilterSegmentOutput,
-} from "./types";
 
 export const listHandler = async ({
   ctx,
@@ -19,77 +11,8 @@ export const listHandler = async ({
   };
   input: TListFilterSegmentsInputSchema;
 }) => {
-  const { tableIdentifier } = input;
-  const userId = ctx.user.id;
-
-  // Get all teams that the user is a member of
-  const userTeamIds = await prisma.membership
-    .findMany({
-      where: {
-        userId,
-        accepted: true,
-      },
-      select: {
-        teamId: true,
-      },
-    })
-    .then((memberships) => memberships.map((m) => m.teamId));
-
-  // Fetch both user-scoped and team-scoped segments
-  const segments = await prisma.filterSegment.findMany({
-    where: {
-      tableIdentifier,
-      OR: [
-        // User-scoped segments created by the current user
-        {
-          scope: "USER",
-          userId,
-        },
-        // Team-scoped segments for teams the user belongs to
-        {
-          scope: "TEAM",
-          teamId: {
-            in: userTeamIds,
-          },
-        },
-      ],
-    },
-    include: {
-      team: {
-        select: {
-          id: true,
-          name: true,
-        },
-      },
-    },
-    orderBy: [
-      { scope: "desc" }, // USER segments first, then TEAM segments
-      { createdAt: "desc" }, // Newest first within each scope
-    ],
+  return await FilterSegmentRepository.get({
+    userId: ctx.user.id,
+    tableIdentifier: input.tableIdentifier,
   });
-
-  const parsedSegments: FilterSegmentOutput[] = segments.map((segment) => ({
-    ...segment,
-    activeFilters: ZActiveFilters.catch([]).parse(segment.activeFilters),
-    sorting: ZSortingState.catch([]).parse(segment.sorting),
-    columnVisibility: ZColumnVisibility.catch({}).parse(segment.columnVisibility),
-    columnSizing: ZColumnSizing.catch({}).parse(segment.columnSizing),
-  }));
-
-  const preference = await prisma.userFilterSegmentPreference.findUnique({
-    where: {
-      userId_tableIdentifier: {
-        userId,
-        tableIdentifier,
-      },
-    },
-    select: {
-      segmentId: true,
-    },
-  });
-
-  return {
-    segments: parsedSegments,
-    preferredSegmentId: preference?.segmentId || null,
-  };
 };
