--- packages/trpc/server/routers/viewer/filterSegments/list.schema.ts
@@ -1,7 +1,0 @@
-import { z } from "zod";
-
-export const ZListFilterSegmentsInputSchema = z.object({
-  tableIdentifier: z.string(),
-});
-
-export type TListFilterSegmentsInputSchema = z.infer<typeof ZListFilterSegmentsInputSchema>;
